dept: refactor common parts of proxy setup for exec and port-forward chore: update callers
