+++ internal/api/v1/application/portforward.go
@@ -13,15 +13,11 @@ package application
 
 import (
 	"net/http"
-	"net/http/httputil"
-	"time"
 
 	"github.com/epinio/epinio/helpers/kubernetes"
 	"github.com/epinio/epinio/internal/application"
 	apierror "github.com/epinio/epinio/pkg/api/core/v1/errors"
 	"github.com/gin-gonic/gin"
-	thekubernetes "k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/rest"
 )
 
 func (hc Controller) PortForward(c *gin.Context) apierror.APIErrors {
@@ -31,11 +27,6 @@ func (hc Controller) PortForward(c *gin.Context) apierror.APIErrors {
 	instanceName := c.Query("instance")
 
 	cluster, err := kubernetes.GetCluster(ctx)
-	if err != nil {
-		return apierror.InternalError(err)
-	}
-
-	clientSetHTTP1, err := kubernetes.GetHTTP1Client(ctx)
 	if err != nil {
 		return apierror.InternalError(err)
 	}
@@ -80,33 +71,14 @@ func (hc Controller) PortForward(c *gin.Context) apierror.APIErrors {
 		podToConnect = podNames[0]
 	}
 
-	forwardRequest(c.Writer, c.Request, podToConnect, namespace, clientSetHTTP1)
-
-	return nil
-}
-
-func forwardRequest(rw http.ResponseWriter, req *http.Request, podName, namespace string, client thekubernetes.Interface) {
 	// https://github.com/kubernetes/kubectl/blob/2acffc93b61e483bd26020df72b9aef64541bd56/pkg/cmd/portforward/portforward.go#L409
-	forwardURL := client.CoreV1().RESTClient().
+	forwardURL := cluster.Kubectl.CoreV1().RESTClient().
 		Post().
 		Resource("pods").
 		Namespace(namespace).
-		Name(podName).
+		Name(podToConnect).
 		SubResource("portforward").
 		URL()
 
-	httpClient := client.CoreV1().RESTClient().(*rest.RESTClient).Client
-	p := httputil.ReverseProxy{
-		Director: func(req *http.Request) {
-			req.URL = forwardURL
-			req.Host = forwardURL.Host
-			// let kube authentication work
-			delete(req.Header, "Cookie")
-			delete(req.Header, "Authorization")
-		},
-		Transport:     httpClient.Transport,
-		FlushInterval: time.Millisecond * 100,
-	}
-
-	p.ServeHTTP(rw, req)
+	return runProxy(ctx, c.Writer, c.Request, forwardURL)
 }
